feat(storybook): add new story for uui-table-row
